+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.1.1] - 2022-08-01
+### Added
+- Support multiple subreddits and 'run all' command.
+- Add 'number of posts' setting which controls the number of posts retrieved for each subreddit.
+
 ## [0.1.0] - 2022-08-01
 ### Added
 - This is the very beginning of roam-reddit. It supports pulling Reddit posts from the subreddits of your choosing into your Roam graph, with some basic configuration options. See the [README](/README.md) for more details.